Set the default stage to production. Deploy from master branch.
We have a `staging` environment defined that runs on a local VMWare
install. Since it's not generally accessible and the only other
environment is production it makes sense to deploy there by default.

I reset the deploy branch to be master now that testing of capistrano is
coming to a close.