Adds an initial stab at LDAP integration
This LDAP authentication piece is about 90% of what we need for this to
be fully functional. It has not been tested at this point. I have
proofed the LDAP authentication code and I know that it does work. It is
the integration with Devise and creating a user that is not complete.

Currently, Users are not valid unless they have a role and a G Number.
We do not currently get either of those pieces from LDAP. This means
that unless we change the database validations, we will require a User
exists in the database in order to log in.