Do not create new records from successful LDAP authentications
If LDAP returns results, we simply want to authenticate the user. We don't want to create new records, as 1. Cerebro should only be used by MLL faculty/students who are in the BlackBoard data import and 2. we would be missing crucial data such as user role and enrollment data. Therefore, simply find the user; if they are not in the database, fail with a custom error message.